minor design fixes; player card now appears in all pages except auth and game
