+++ web/home.js
@@ -8,41 +8,10 @@ import {renderPongGame} from './pong_game.js'
 	/*	PAGE RENDERING	*/
 function	renderPage()
 {
-	let	transcendenceApp = document.getElementById("appContainer");
+	let	transcendenceApp = document.getElementById("ctrlsRow");
 
 	transcendenceApp.innerHTML = `
-			<div class="row text-center d-flex justify-content-center">
-				<div class="col-12 col-lg-2 mt-lg-4 border rounded-circle">
-					<!--PROFILE PIC-->
-<!--
-					<img src="" class="img-fluid rounded-circle" alt="User Profile Picture">
--->
-				</div>
-				<div class="col-12 col-lg-8 mx-lg-2 mt-4 border rounded">
-					<div class="row h-auto">
-						<div class="col d-flex justify-content-end align-items-start">
-							<!--EDIT PROFILE BTN (CHG DISPLAY NAME, CHG MOTTO, CHG PFP)-->
-							<button type="button" class="btn btn-sm btn-outline-light mx-2 my-2"><i class="bi-pencil-fill"></i></button>
-							<!--SETTINGS BTN (CHG PASSWD, DELETE ACCOUNT)-->
-							<button type="button" class="btn btn-sm btn-outline-light my-2"><i class="bi-gear-fill"></i></button>
-						</div>
-					</div>
-					<div class="flex-row flex-fill align-middle">
-						<div class="col py-auto d-flex justify-content-center">
-							<!--USERNAME-->
-							<h1 id="userNameDisplay" class="display-1"><a href="#" class="link-light link-underline link-underline-opacity-0 link-opacity-75-hover">$USER</a></h1>
-						</div>
-					</div>
-					<div class="row d-flex align-items-end">
-						<div class="col">
-							<!--MOTTO-->
-							<p class="fst-italic">"Some days you are the pidgeon, some days you are the statue. Today it's clearly statue day."</p>
-						</div>
-					</div>
-				</div>
-			</div>
-			<div class="row text-center justify-content-center">
-				<div class="col-12 col-lg-5 mx-lg-1 mt-4 border">
+				<div class="col-12 col-lg-5 mx-lg-1 my-2 border">
 					<div class="row">
 						<div class="col-12">
 							<h1 class="mt-4">PLAY PONG</h1>
@@ -55,7 +24,7 @@ function	renderPage()
 						</div>
 					</div>
 				</div>
-				<div class="col-12 col-lg-5 mx-lg-1 mt-4 px-5 border">
+				<div class="col-12 col-lg-5 mx-lg-1 my-2 px-5 border">
 					<h1 class="mt-4">SETTINGS</h1>
 					<div class="row my-1 justify-content-center">
 						<div class="col col-lg-10 mt-4">
@@ -107,7 +76,6 @@ function	renderPage()
 						</div>
 					</div>
 				</div>
-			</div>
 	`;
 }
 
@@ -125,7 +93,7 @@ function	selectSettings()
 function	setupEventHandlers()
 {
 
-	document.getElementById("userNameDisplay").addEventListener("click", ()=> renderProfile() );
+
 
 	document.getElementById("btnFriendlyMatch").addEventListener("click", ()=> renderPongGame() );
 	document.getElementById("btnTournament").addEventListener("click", ()=> alert("This feature has not been implemented yet.") ) ;
